Move cache-related methods to new Cache class
CO-955

The Cache class manages cached data. Move every method on LocationUtil
that needed the cacheDirectory variable onto Cache instead.
