+++ src/main/groovy/edu/oregonstate/mist/locations/db/CampusMapDAO.groovy
@@ -1,6 +1,6 @@
 package edu.oregonstate.mist.locations.db
 
-import edu.oregonstate.mist.locations.LocationUtil
+import edu.oregonstate.mist.locations.Cache
 import edu.oregonstate.mist.locations.core.CampusMapLocation
 import groovy.json.JsonSlurper
 
@@ -10,12 +10,12 @@ class CampusMapDAO {
 
     private final String campusMapJsonOut
 
-    private final LocationUtil locationUtil
+    private final Cache cache
 
-    CampusMapDAO(Map<String, String> locationConfiguration, LocationUtil locationUtil) {
+    CampusMapDAO(Map<String, String> locationConfiguration, Cache cache) {
         campusMapJsonUrl = locationConfiguration.get("campusMapHttpData")
         campusMapJsonOut = locationConfiguration.get("campusmapJsonOut")
-        this.locationUtil = locationUtil
+        this.cache = cache
     }
 
     /**
@@ -39,7 +39,7 @@ class CampusMapDAO {
      */
     private List<CampusMapLocation> getCampusMapJson() {
         def jsonSlurper = new JsonSlurper()
-        String campusMapData = locationUtil.getDataFromUrlOrCache(
+        String campusMapData = cache.getDataFromUrlOrCache(
                 campusMapJsonUrl,
                 campusMapJsonOut)
